fix: 타입 변경

+++ group-investment/src/main/java/com/example/group_investment/team/dto/DetailPendingTeamResponse.java
@@ -2,7 +2,6 @@ package com.example.group_investment.team.dto;
 
 import com.example.group_investment.enums.Category;
 import com.example.group_investment.enums.RulePeriod;
-import com.example.group_investment.enums.RuleType;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -17,20 +16,20 @@ public class DetailPendingTeamResponse {
     private Category category;
     private LocalDateTime startAt;
     private LocalDateTime endAt;
-    private float prdyVrssRt;
+    private double prdyVrssRt;
     private int urgentTradeUpvotes;
     private int tradeUpvotes;
     private int depositAmt;
     private RulePeriod period;
     private Date payDate;
-    private float maxLossRt;
-    private float maxProfitRt;
+    private double maxLossRt;
+    private double maxProfitRt;
     private int invitedMembers;
     private int isLeader;
     private int isParticipating;
 
     @Builder
-    public DetailPendingTeamResponse(String name, int baseAmt, int headCount, Category category, LocalDateTime startAt, LocalDateTime endAt, float prdyVrssRt, int urgentTradeUpvotes, int tradeUpvotes, int depositAmt, RulePeriod period, Date payDate, float maxLossRt, float maxProfitRt, int invitedMembers, int isLeader, int isParticipating) {
+    public DetailPendingTeamResponse(String name, int baseAmt, int headCount, Category category, LocalDateTime startAt, LocalDateTime endAt, double prdyVrssRt, int urgentTradeUpvotes, int tradeUpvotes, int depositAmt, RulePeriod period, Date payDate, double maxLossRt, double maxProfitRt, int invitedMembers, int isLeader, int isParticipating) {
         this.name = name;
         this.baseAmt = baseAmt;
         this.headCount = headCount;
@@ -49,6 +48,7 @@ public class DetailPendingTeamResponse {
         this.isLeader = isLeader;
         this.isParticipating = isParticipating;
     }
+
     public DetailPendingTeamResponse() {
 
     }
